Overhaul account format
When a purchase is made, its amount is
- added to the real account Expenses:$Category
- deducted from the real account Liabilities:People:$Purchaser

And, for each purchasee, the amount is divided by the number of
purchasees, and that new amount is
- added to the virtual account (Assets:People:$Purchasee)

When a payment is made, its amount is
- added to the real account Liabilities:People:$Payee
- deducted from the real account Income:People:$Payer
- deducted from the virtual account (Assets:People:$Payer)

This allows us to track real purchases using the `--real` flag, which
shows us only transactions which have actually happened in the real
world.

Also, for each person, we can find the amount which they owe in total as
the sum of their liabilities (amount owed, a negative value) and assets
(amount they owe, a positive value). Note, the sign of the value is from
the perspective of the "network", not from any individual.

Therefore, we to find the balance of a person,
```
ledger -f sheet.ledger balance Assets:People:$Person Liabilities:People:$Person
```