Change Asset tracking
When a purchase is made, the amount is
- Added to Expenses:$Category
- Deducted from Liabilities:People:$Purchaser
- Added to (Assets:People:$Purchasee) divided by the number of
  purchasees, for each purchasee.

This is exactly the same as it was before.

When a payment is made, the amount is
- Added to Liabilities:People:$Payee
- Deducted from Income:People:$Payer
- NOT deducted from (Assets:People:$Payer)

This differs from the previous version in that no value is deducted from
(Assets:People:$Payee). This is because the real balance is then correct
across all accounts (Expenses, Liabilities, and Income balance), and
amount owed by any person can be retrieved by querying all accounts for
that person.